created useImpactPlanManager custom hook to abstract handler functions & state from impactplan settings
